Fixed a few bugs with iteration and COW
