+++ Sources/SAKBase/Collections/MutableContainerArray.swift
@@ -37,65 +37,6 @@ extension MutableContainerArray: Copyable {
 	}
 }
 
-extension MutableContainerArray: Collection {
-
-	public typealias Iterator = IndexingIterator<MutableContainerArray>
-
-	public typealias Element = Container
-
-	public typealias Index = Int
-
-	@inlinable
-	public var isEmpty: Bool {
-		internalArray.isEmpty
-	}
-
-	@inlinable
-	public var count: Int {
-		internalArray.count
-	}
-
-	@inlinable
-	public var first: Element? {
-		internalArray.first
-	}
-
-	@inlinable
-	public var startIndex: Index {
-		internalArray.startIndex
-	}
-
-	@inlinable
-	public var endIndex: Index {
-		internalArray.endIndex
-	}
-
-	@inlinable
-	public subscript(_ position: Index) -> Element {
-		get {
-			internalArray[position]
-		}
-		set {
-			internalArray[position] = newValue
-		}
-	}
-
-	@inlinable
-	public subscript(_ position: Index) -> ElementContained {
-		get {
-			internalArray[position].wrappedValue
-		}
-		set {
-			internalArray[position].wrappedValue = newValue
-		}
-	}
-
-	@inlinable
-	public func index(after index: Index) -> Int {
-		internalArray.index(after: index)
-	}
-}
-
 // MARK: Element Methods
 
 extension MutableContainerArray: ExpressibleByArrayLiteral {
@@ -113,13 +54,6 @@ extension MutableContainerArray: ExpressibleByArrayLiteral {
 		// Swift gets confused if we try to call the Sequence initializer.
 		self.init()
 		self.internalArray = ContiguousArray(elements.map { Container($0) })
-	}
-}
-
-extension MutableContainerArray: BidirectionalCollection {
-	@inlinable
-	public func index(before index: Int) -> Int {
-		internalArray.index(before: index)
 	}
 }
 
@@ -147,9 +81,11 @@ extension MutableContainerArray: RangeReplaceableCollection {
 		let inScopeRange = subrange.clamped(to: internalArray.indices)
 		let outOfScopeRange = subrange.clamped(to: internalArray.endIndex..<subrange.endIndex)
 
-		for (referenceElement, newElement) in zip(internalArray[inScopeRange], newElements) {
-			// For each inscope index we just update the given range
-			referenceElement.wrappedValue = newElement
+		internalArray.withUnsafeBufferPointer { unsafeArray in
+			for (referenceElement, newElement) in zip(unsafeArray[inScopeRange], newElements) {
+				// For each inscope index we just update the given range
+				referenceElement.wrappedValue = newElement
+			}
 		}
 
 		let appendingElements = newElements[newElements.index(newElements.startIndex, offsetBy: inScopeRange.count)...]
